[Hd] Moved ID render decoding to HdxIntersector It belongs in HdxIntersector, and not HdxRenderSetupTask.
(Internal change: 1830067)
